+++ imod/mf6/gwfgwf.py
@@ -92,6 +92,7 @@ class GWFGWF(ExchangeBase):
         top: Optional[GridDataArray] = None,
         bottom: Optional[GridDataArray] = None,
     ) -> Self:
+        """This package cannot be clipped by a bounding box."""
         raise NotImplementedError("this package cannot be clipped")
 
     def _render(