+++ imod/mf6/gwtgwt.py
@@ -88,6 +88,7 @@ class GWTGWT(ExchangeBase):
         top: Optional[GridDataArray] = None,
         bottom: Optional[GridDataArray] = None,
     ) -> Self:
+        """This package cannot be clipped by a bounding box."""
         raise NotImplementedError("this package cannot be clipped")
 
     def _render(